requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
django-cors-headers
gspread
google-auth
reportlab
django-storages[boto3]
python-dotenv

# Production dependencies for Cloud Run
gunicorn==21.2.0
whitenoise==6.6.0
google-cloud-storage==2.14.0
